# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  message: string;
}

/**
 * A numbered cycling node (knooppunt).
 */
export interface NetworkNode {
  /** Stable OSM node id. */
  id: string;
  /** The knooppunt number shown on signage (e.g. "34"). */
  ref: string;
  lat: number;
  lon: number;
}

/**
 * A network path geometry connecting points in the node network.
 */
export interface NetworkSegment {
  id: string;
  /** Ordered [lon, lat] pairs describing the segment line. */
  coordinates: number[][];
}

export interface NetworkData {
  nodes: NetworkNode[];
  segments: NetworkSegment[];
  /** True if the area was too large and results were limited. */
  truncated: boolean;
}

export interface RouteRequestNode {
  id: string;
  ref: string;
  lat: number;
  lon: number;
}

export interface RouteRequest {
  /** Ordered list of selected nodes to route through. */
  nodes: RouteRequestNode[];
}

/**
 * One leg of the route between two consecutive nodes.
 */
export interface RouteLeg {
  fromRef: string;
  toRef: string;
  distanceMeters: number;
  coordinates: number[][];
}

export interface RoutePlan {
  /** The ordered knooppunt numbers along the route. */
  nodeRefs: string[];
  /** Full route geometry as ordered [lon, lat] pairs. */
  coordinates: number[][];
  distanceMeters: number;
  legs: RouteLeg[];
}

export interface Region {
  id: string;
  name: string;
  /** ISO country code (NL or BE). */
  country: string;
  lat: number;
  lon: number;
  zoom: number;
}

export type GetNetworkParams = {
/**
 * Bounding box as minLon,minLat,maxLon,maxLat (WGS84).
 */
bbox: string;
};