# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /route:
    post:
      operationId: planRoute
      tags: [routing]
      summary: Plan a route along the node network
      description: >
        Given an ordered list of selected nodes, computes the route that
        follows the cycling network segments between each consecutive pair of
        nodes, returning the full geometry, the ordered node numbers, and the
        total distance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Planned route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutePlan"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: No connecting path found between the selected nodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Upstream data source error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    RouteRequestNode:
      type: object
      properties:
        id:
          type: string
        ref:
          type: string
        lat:
          type: number
        lon:
          type: number
      required:
        - id
        - ref
        - lat
        - lon
    RouteRequest:
      type: object
      properties:
        nodes:
          type: array
          description: Ordered list of selected nodes to route through.
          items:
            $ref: "#/components/schemas/RouteRequestNode"
      required:
        - nodes
    RouteLeg:
      type: object
      description: One leg of the route between two consecutive nodes.
      properties:
        fromRef:
          type: string
        toRef:
          type: string
        distanceMeters:
          type: number
        coordinates:
          type: array
          items:
            type: array
            items:
              type: number
      required:
        - fromRef
        - toRef
        - distanceMeters
        - coordinates
    RoutePlan:
      type: object
      properties:
        nodeRefs:
          type: array
          description: The ordered knooppunt numbers along the route.
          items:
            type: string
        coordinates:
          type: array
          description: "Full route geometry as ordered [lon, lat] pairs."
          items:
            type: array
            items:
              type: number
        distanceMeters:
          type: number
        legs:
          type: array
          items:
            $ref: "#/components/schemas/RouteLeg"
      required:
        - nodeRefs
        - coordinates
        - distanceMeters
        - legs
...