# lib/api-zod/src/generated/types/routeLeg.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * One leg of the route between two consecutive nodes.
 */
export interface RouteLeg {
  fromRef: string;
  toRef: string;
  distanceMeters: number;
  coordinates: number[][];
}