# lib/api-zod/src/generated/types/routePlan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouteLeg } from './routeLeg';

export interface RoutePlan {
  /** The ordered knooppunt numbers along the route. */
  nodeRefs: string[];
  /** Full route geometry as ordered [lon, lat] pairs. */
  coordinates: number[][];
  distanceMeters: number;
  legs: RouteLeg[];
}