# lib/api-zod/src/generated/types/routeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouteRequestNode } from './routeRequestNode';

export interface RouteRequest {
  /** Ordered list of selected nodes to route through. */
  nodes: RouteRequestNode[];
}